requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-multipart
openai
python-dotenv